+++ CHANGES.md
@@ -9,6 +9,7 @@ Release Notes.
 * Support propagation the excluding namespaces in the access log to the backend.
 * Add `pprof` module for observe self.
 * Add detect process from `CRI-O` container in Kubernetes.
+* Introduce `MonitorFilter` into access log module. 
 
 #### Bug Fixes
 * Fixed the issue where `conntrack` could not find the Reply IP in the access log module.